+++ Thunderstore/CHANGELOG.md
@@ -4,6 +4,7 @@
 - Changed BatteryDrainPercent to an integer range from 0 to 100
 - Improving init and update performance
 - Fix issue where disconnecting players wouldn't drop some items
+- Fix issue where dead players wouldn't drop some items
 - Fix incorrect default value on ResetCooldownOnOrbit
 - Fix issue where changing server host synced incorrect settings