+++ Thunderstore/CHANGELOG.md
@@ -1,11 +1,14 @@
+## Version 1.1.1
+
+- Fix issue where dead players wouldn't drop some items
+- Fix issue where reconnecting players would always drop all items when teleporting
+
 ## Version 1.1.0
 
 - Using CSync for better network communication
 - Changed BatteryDrainPercent to an integer range from 0 to 100
 - Improving init and update performance
 - Fix issue where disconnecting players wouldn't drop some items
-- Fix issue where reconnecting players would always drop all items when teleporting
-- Fix issue where dead players wouldn't drop some items
 - Fix incorrect default value on ResetCooldownOnOrbit
 - Fix issue where changing server host synced incorrect settings